Сorrection of the specified errors
