+++ TestWeather/TestWeather/Application/SceneDelegate.swift
@@ -20,7 +20,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         
         window = UIWindow(frame: windowScene.coordinateSpace.bounds)
         window?.windowScene = windowScene
-        window?.rootViewController = UINavigationController(rootViewController: ForecastWeathersViewController())
+        window?.rootViewController = UINavigationController(rootViewController: ForecastWeatherViewController())
         window?.makeKeyAndVisible()
     }
